backend/requirements.txt
mysql-connector-python==8.4.0
PyJWT==2.12.1
